+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * fix odd? even? on non integers
 * fix object literals with null value [#264](https://github.com/jcubic/lips/issues/264)
 * fix version and date in executable [#261](https://github.com/jcubic/lips/issues/261)
+* fix error on extra close parenthesis [#263](https://github.com/jcubic/lips/issues/263)
 
 ## 1.0.0-beta.17
 ### Breaking